feat(dashboard): add ContentModel & ContentModelContext
